+++ sdk/packages/sdk/src/configs/ChainConfigService.ts
@@ -187,8 +187,8 @@ export class ChainConfigService {
 		return this.getConfig(chain)?.addresses.EntryPointV08!
 	}
 
-	getCirclePaymasterV08Address(chain: string): HexString | undefined {
-		return this.getConfig(chain)?.addresses.CirclePaymasterV08 as HexString | undefined
+	getSimplexPaymasterAddress(chain: string): HexString | undefined {
+		return this.getConfig(chain)?.addresses.SimplexPaymaster as HexString | undefined
 	}
 
 	getHyperbridgeAddress(): string {
