+++ sdk/packages/simplex/src/services/wallet/types.ts
@@ -82,7 +82,7 @@ export interface SigningAccount extends SdkSigningAccount {
 	account: Account
 	mode: "privateKey" | "mpcVault" | "turnkey"
 	/**
-	 * Signs an EIP-712 typed-data payload (e.g. an EIP-2612 USDC permit for the Circle paymaster).
+	 * Signs an EIP-712 typed-data payload (e.g. an EIP-2612 USDC permit for the SimplexPaymaster).
 	 * The shape of `typedData` matches viem's `TypedDataDefinition`.
 	 * MPC adapter must JSON.stringify before delegating to MpcVaultService.signTypedData.
 	 */
